backend/requirements.txt
django==4.2.1
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-import-export==3.2.0
django-mptt==0.14.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.11
pillow==9.5.0
python-decouple==3.8
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
drf-yasg[validation]==1.21.6
coreapi==2.3.3
django-filter==23.2
